client: rename protoutils to proto_utils
