+++ client/zeth/core/wallet.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 from zeth.core.zeth_address import ZethAddressPriv
 from zeth.core.mixer_client import \
     MixOutputEvents, compute_nullifier, compute_commitment, receive_note
-from zeth.core.protoutils import zeth_note_from_json_dict, zeth_note_to_json_dict
+from zeth.core.proto_utils import zeth_note_from_json_dict, zeth_note_to_json_dict
 from zeth.core.constants import ZETH_MERKLE_TREE_DEPTH
 from zeth.core.pairing import PairingParameters
 from zeth.core.merkle_tree import ITreeHash, PersistentMerkleTree
